Move code into testcode since it should only be exposed to tests and not users
